# src/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "v1.0.0",
    "title": "LCDWallet API Documentation",
    "description": "Documentation for the LCDWallet API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/blocks/mine": {
      "post": {
        "tags": [
          "Blocks"
        ],
        "summary": "Mine a new block",
        "description": "Creates and adds a new block to the blockchain if the validator is authorized.",
        "parameters": [
          {
            "name": "validatorAddress",
            "in": "query",
            "description": "Public key of the validator",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The newly mined block.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "example": {
                      "type": "string",
                      "example": "Invalid validator address"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "example": {
                      "type": "string",
                      "example": "Invalid validator address"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "example": {
                      "type": "string",
                      "example": "An error occurred while mining the block"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "string"
                    },
                    "example": {
                      "type": "string",
                      "example": "An error occurred while mining the block"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/blocks/all": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "Get All Blocks",
        "description": "Retrieves all blocks in the blockchain, starting from the genesis block.",
        "responses": {
          "200": {
            "description": "An array containing all blocks in the blockchain.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "array"
                    },
                    "items": {
                      "$ref": "#/components/schemas/Block"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "type": {
                      "type": "string",
                      "example": "array"
                    },
                    "items": {
                      "$ref": "#/components/schemas/Block"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. An unexpected error occurred while fetching the blocks."
          }
        }
      }
    },
    "/blocks/index/{index}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "Get Block By Index",
        "description": "Retrieves a block by its index in the blockchain.",
        "parameters": [
          {
            "name": "index",
            "in": "path",
            "description": "Index of the block to retrieve.",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested block.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "404": {
            "description": "Block not found."
          }
        }
      }
    },
    "/blocks/hash/{hash}": {
      "get": {
        "tags": [
          "Blocks"
        ],
        "summary": "Get Block By Hash",
        "description": "Retrieves a block by its hash value.",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "description": "Hash of the block to retrieve.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested block.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Block"
                }
              }
            }
          },
          "404": {
            "description": "Block not found."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Block": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "index": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "hash": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "previousHash": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "timestamp": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  }
                }
              },
              "transactions": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "items": {
                    "xml": {
                      "name": "transaction"
                    },
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              },
              "validator": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "signature": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "Block"
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Unique identifier of the transaction (usually a hash)."
                  }
                }
              },
              "txIns": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "description": {
                    "type": "string",
                    "example": "Array of transaction inputs (TxIn)."
                  },
                  "items": {
                    "xml": {
                      "name": "txin"
                    },
                    "$ref": "#/components/schemas/TxIn"
                  }
                }
              },
              "txOuts": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "array"
                  },
                  "description": {
                    "type": "string",
                    "example": "Array of transaction outputs (TxOut)."
                  },
                  "items": {
                    "xml": {
                      "name": "txout"
                    },
                    "$ref": "#/components/schemas/TxOut"
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "Transaction"
        }
      },
      "TxIn": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "txOutId": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Transaction ID of the UTXO being spent."
                  }
                }
              },
              "txOutIndex": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  },
                  "description": {
                    "type": "string",
                    "example": "Output index of the UTXO being spent."
                  }
                }
              },
              "signature": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Signature proving ownership of the UTXO."
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "TxIn"
        }
      },
      "TxOut": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "address": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "description": {
                    "type": "string",
                    "example": "Address of the recipient."
                  }
                }
              },
              "amount": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "number"
                  },
                  "description": {
                    "type": "string",
                    "example": "Amount of cryptocurrency being sent."
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "TxOut"
        }
      }
    }
  }
}